Moved IconBox to own file
