+++ frontend/src/components/Buttons/ActionButton.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import BaseButton from "./BaseButton";
+import { IconBox } from "./IconBox";
 
 export type ActionButtonProps = {
   variant: "primary" | "secondary" | "terniary";
@@ -51,23 +52,5 @@ export default function ActionButton({
       {children}
       <IconBox small={small}>{iconRight}</IconBox>
     </BaseButton>
-  );
-}
-
-export function IconBox({
-  children,
-  small,
-}: {
-  children: React.ReactNode;
-  small?: boolean;
-}) {
-  const size = small ? "h-4 w-4" : "h-6 w-6";
-
-  if (!children) {
-    return null;
-  }
-
-  return (
-    <div className={`flex items-center justify-center ${size}`}>{children}</div>
   );
 }
